Move childComponent to a different component and import it
